[minimum-amount-aud] Merge contributionAmounts, contributionTypes and contributions all in one helper and updated its usage,
